+++ src/fairchem/core/components/calculate/simulation_tools/trajectory.py
@@ -182,87 +182,6 @@ class TrajectoryFrame:
             spin=spin,
         )
 
-    def to_atoms(self) -> Atoms:
-        """
-        Reconstruct an ASE Atoms object from this frame.
-
-        Returns:
-            Atoms object with positions, cell, tags, constraints, and info.
-        """
-        from ase import Atoms as _Atoms
-
-        atoms = _Atoms(
-            numbers=self.atomic_numbers,
-            positions=self.positions,
-            cell=self.cell,
-            pbc=self.pbc,
-        )
-        if self.tags is not None:
-            atoms.set_tags(self.tags)
-        if self.velocities is not None:
-            atoms.set_velocities(self.velocities)
-        if self.fixed is not None:
-            from ase.constraints import FixAtoms
-
-            atoms.constraints = [FixAtoms(indices=np.where(self.fixed)[0])]
-        if self.sid is not None:
-            atoms.info["sid"] = self.sid
-        if self.charge is not None:
-            atoms.info["charge"] = self.charge
-        if self.spin is not None:
-            atoms.info["spin"] = self.spin
-        return atoms
-
-    @classmethod
-    def from_dict(cls, d: dict) -> TrajectoryFrame:
-        """
-        Create a TrajectoryFrame from a dictionary (e.g. a parquet row).
-
-        Args:
-            d: Dictionary with trajectory frame data.
-
-        Returns:
-            TrajectoryFrame populated from the dictionary.
-        """
-
-        def _to_array(val, dtype=float):
-            """
-            Convert a value to a numpy array, handling parquet-style
-            nested lists (list of arrays) that np.array() can't auto-stack.
-            """
-            if isinstance(val, np.ndarray) and val.dtype == object:
-                return np.stack(val).astype(dtype)
-            return np.asarray(val, dtype=dtype)
-
-        return cls(
-            step=d["step"],
-            atomic_numbers=_to_array(d["atomic_numbers"], dtype=int),
-            positions=_to_array(d["positions"]),
-            cell=_to_array(d["cell"]),
-            pbc=_to_array(d["pbc"], dtype=bool),
-            sid=d.get("sid"),
-            time=d.get("time"),
-            velocities=(
-                _to_array(d["velocities"]) if d.get("velocities") is not None else None
-            ),
-            energy=d.get("energy"),
-            forces=(_to_array(d["forces"]) if d.get("forces") is not None else None),
-            stress=(_to_array(d["stress"]) if d.get("stress") is not None else None),
-            temperature=d.get("temperature"),
-            kinetic_energy=d.get("kinetic_energy"),
-            pressure=d.get("pressure"),
-            tags=(
-                _to_array(d["tags"], dtype=int) if d.get("tags") is not None else None
-            ),
-            fixed=(
-                _to_array(d["fixed"], dtype=bool)
-                if d.get("fixed") is not None
-                else None
-            ),
-            charge=d.get("charge"),
-            spin=d.get("spin"),
-        )
-
 
 class ParquetTrajectoryWriter:
     """
